Update project with new Validator class and City class
This commit introduces a new Validator class used to centralize logic for validating different types of inputs. Changes in Human class reflect the use of this new Validator. The commit also adds a new City class with corresponding fields and methods. Other minor changes include updated paths in the ChangelistManager and tweaks to workspace settings.